start blocks page
