+++ frontend/src/dummyApi.js
@@ -1,11 +1,12 @@
 const blockchain = require('./blockchain.json');
 
 const api = () => {
-    const latestBlx = blockchain.chain.reverse().slice(0, 5);
+    const blocks = blockchain.chain.reverse();
+    const latestBlx = blocks.slice(0, 5);
     let latestTxs = [];
-    latestBlx.forEach(b => latestTxs.push(...b.txs));
-    latestTxs = latestTxs.reverse().slice(0, 5);
-    return { latestBlx, latestTxs };  
+    latestBlx.forEach(b => latestTxs.push(...b.txs.reverse()));
+    latestTxs = latestTxs.slice(0, 5);
+    return { blocks, latestBlx, latestTxs };  
 };
 
 module.exports = api;
